fix: change layout and direction files/upd readme/change layout and directories
