+++ README.md
@@ -1,1 +1,14 @@
-This repository created for test task from pizza-fabric
+## ТЗ : Необходимо разработать одностраничное приложение для возможности просмотра и редактирования данных сотрудников компании. Внешний вид приложения зависит от вас. Верстка должна быть адаптивной и корректно отображаться на разных устройствах. Стартовый набор данных для приложения находится в файле employees.json.
+
+### Основной стек: Eslint, Prettier, TS, React, Redux, React-hook-form, Scss, Jest, Vite, React-router-dom 
+### Что бы запустить приложение: 
+- Скачайте репозиторий
+- Установите зависимости (npm i)
+- Для запуска в дев режиме (npm run dev)
+- Для запуска тестов (npm run test)
+
+Несколько скриншотов из приложения:
+![image](https://github.com/user-attachments/assets/df062806-4d07-402e-b79c-c6c3ac49195a)
+![image](https://github.com/user-attachments/assets/62232370-c0f0-46db-b6ba-10b122cb1df6)
+![image](https://github.com/user-attachments/assets/dad713e8-1796-48e1-a912-8dc139388398)
+
