+++ src/components/filterContainer/filterContainer.scss
@@ -1,4 +1,4 @@
-@use "../../App.scss" as app;
+@use "../../mixin.scss" as app;
 
 .sort-section-select {
   @include app.flex-positioning(center, center, column);
