+++ src/components/workers/workerForm.scss
@@ -1,4 +1,4 @@
-@use "../../App.scss" as app;
+@use "../../mixin.scss" as app;
 
 .workerForm-form {
   @include app.flex-positioning(center, center, column);
