+++ src/mixin.scss
@@ -1,0 +1,28 @@
+$background-color: #6a6a6a;
+$text-color: #000000;
+$container-color: #d9d9d9;
+$second-color: #f0f0f0;
+$hover-color: #f9fffa;
+
+@mixin position-center {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+@mixin full-size {
+  width: 100%;
+  height: 100%;
+}
+
+@mixin flex-positioning(
+  $justify-content: center,
+  $align-items: center,
+  $flex-direction: row
+) {
+  display: flex;
+  justify-content: $justify-content;
+  align-items: $align-items;
+  flex-direction: $flex-direction;
+}
